started vertex lightning

+++ QtGCanvas/qtgdrawboilerplate.cpp
@@ -3,24 +3,6 @@
 namespace drawers {
 
 
-
-
-
-//void QtGDrawBoilerPlate::_enable_vbo()
-//{
-//    canvas->glBindBuffer(GL_ARRAY_BUFFER, vboIds[0]);
-//    canvas->glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, vboIds[1]);
-
-//    _shared_enable_attrs();
-//}
-
-//void QtGDrawBoilerPlate::_enable_vbo(DGL_VBO *shared_vbo)
-//{
-//    shared_vbo->bind();
-//    _shared_enable_attrs();
-//}
-
-
 static inline char* _gl_err2str(GLenum err){
     switch (err) {
     case GL_INVALID_ENUM:
@@ -53,7 +35,7 @@ void QtGDrawBoilerPlate::_shared_enable_attrs()
 {
     quintptr offset = 0;
 
-    for (int i = 0; i < num_vbo_attribs; i++){
+    for (size_t i = 0; i < num_vbo_attribs; i++){
         vbo_locdef_t* tld = vbo_attribs+i;
         _bp_canvas->glEnableVertexAttribArray(tld->a_loc);
         _bp_canvas->glVertexAttribPointer(tld->a_loc, tld->asize, tld->atype, GL_FALSE, this->instance_stride, (const void *)offset);
@@ -66,12 +48,11 @@ bool QtGDrawBoilerPlate::init()
     this->_bp_canvas->set_shader(this->shader, this);
     QGLShaderProgram* shdp = this->shader->program();
 
-
     this->_init_vbo_attribs();
     this->_init_vbo_data();
 
     instance_stride = 0;
-    for(int i = 0; i < this->num_vbo_attribs; i++){
+    for(size_t i = 0; i < this->num_vbo_attribs; i++){
         (this->vbo_attribs+i)->a_loc = shdp->attributeLocation((this->vbo_attribs+i)->aname);
         instance_stride+= (this->vbo_attribs+i)->cstep;
     }
@@ -86,9 +67,7 @@ bool QtGDrawBoilerPlate::init()
 
 void QtGDrawBoilerPlate::deactivate()
 {
-    //    canvas->glBindBuffer(GL_ARRAY_BUFFER, 0);
-    //    canvas->glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, 0);
-    for(int i =0; i < this->num_vbo_attribs; i++){
+    for(size_t i =0; i < this->num_vbo_attribs; i++){
         _bp_canvas->glDisableVertexAttribArray((vbo_attribs+i)->a_loc);
     }
     shader->program()->release();
